3.0.10

+++ src/methods/bigInt.method.ts
@@ -1,3 +1,3 @@
-export function BigIntMethod(argument: unknown): argument is BigInteger {
+export function BigIntMethod(argument: unknown): argument is bigint {
   return typeof argument === 'bigint';
 }
